move: Menu, MenuItem 클래스 domain.menu 패키지로 이동
