+++ src/main/kotlin/christmas/domain/menu/MenuItem.kt
@@ -1,4 +1,6 @@
-package christmas.domain
+package christmas.domain.menu
+
+import christmas.domain.Error
 
 interface MenuItem {
